+++ library/src/main/java/de/mrapp/android/tabswitcher/gesture/SwipeGestureEventHandler.java
@@ -15,11 +15,11 @@ package de.mrapp.android.tabswitcher.gesture;
 
 import android.content.res.Resources;
 import android.graphics.RectF;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.MotionEvent;
 import android.view.ViewConfiguration;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import de.mrapp.android.tabswitcher.Layout;
 import de.mrapp.android.tabswitcher.R;
 import de.mrapp.android.tabswitcher.TabSwitcher;